requirements.txt
dj-database-url==0.4.2
django-bootstrap3==8.2.3
django-extensions==1.7.9
django-hamlpy==1.1
django-nose==1.4.4
Django==1.11.1
gunicorn==19.7.1
ipython==6.0.0
psycopg2==2.7.1
python-dotenv==0.6.4
whitenoise==3.3.0
